% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Repeated CV - GLM}
\usage{
cv(
  expr,
  df,
  family,
  thresh = FALSE,
  statistic = "Accuracy",
  number = 10,
  repeats = 10
)
}
\arguments{
\item{expr}{GLM formula.}

\item{df}{Data.frame.}

\item{family}{Object : a description of the error distribution.}

\item{thresh}{If TRUE, an optimal threshold is found based on the 'statistic' criterion.}

\item{statistic}{A character indicating which statistic to maximize. Default is "Accuracy". You can use c('Sensitivity', 'Specificity', 'Pos Pred Value', 'Neg Pred Value', 'Precision', 'Accuracy', 'Balanced Accuracy', 'F1', 'Detection Rate', 'Detection Prevalence', 'Prevalence', 'Recall', 'Kappa', 'J', 'Dist').}

\item{number}{The number of folds.}

\item{repeats}{The number of complete sets of folds to compute.}
}
\value{
Different accuracy measures.
}
\description{
Repeated CV - GLM
}
