% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data_standarize}
\alias{data_standarize}
\title{Impute missing values}
\usage{
data_standarize(Data, method = "kalman", variables = NA)
}
\arguments{
\item{Data}{data.frame: data frame of target variable, exogenous variables, and observed date (named 'date')}

\item{method}{string: select which method to use from the imputeTS package; 'interpolation', 'kalman', 'locf', 'ma', 'mean', 'random', 'remove','replace', 'seadec', 'seasplit'}

\item{variables}{string: vector of variables to standardize, default is all but 'date' column}
}
\value{
data.frame with missing data imputed
}
\description{
A function to impute missing values. Is used as a data preparation helper function and is called internally
by forecast_univariate, forecast_multivariate, and forecast_combine.
}
