% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{dimension_reduction}
\alias{dimension_reduction}
\title{Dimension reduction via factors}
\usage{
dimension_reduction(
  Data,
  forecast.dates,
  target,
  method,
  ncomp,
  standardize = TRUE,
  rolling.window = NA,
  freq = NA,
  horizon = 0
)
}
\arguments{
\item{Data}{data.frame: data frame of target variable, exogenous variables, and observed date (named 'date')}

\item{forecast.dates}{date: dates forecasts are created}

\item{target}{string: column name in \code{Data} of variable to forecast}

\item{method}{string or vector: methods to use; 'pls', ''pc}

\item{ncomp}{int: number of factors to create}

\item{standardize}{boolean: normalize variables (mean zero, variance one) before estimating factors}

\item{rolling.window}{int: size of rolling window, NA if expanding window is used}

\item{freq}{string: time series frequency; day, week, month, quarter, year; only needed for rolling window factors}

\item{horizon}{int: horizon target will be forecasted (for pls)}
}
\value{
data.frame with a date column and one column per forecast method selected
}
\description{
A function to estimate out-of-sample principal components or partial least square scores.
}
