% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{forecast_date}
\alias{forecast_date}
\title{Set forecasted date}
\usage{
forecast_date(forecast.date, horizon, freq)
}
\arguments{
\item{forecast.date}{date: date forecast was made}

\item{horizon}{int: priods ahead of forecast}

\item{freq}{string: time series frequency; day, week, month, quarter, year; only needed for rolling window factors}
}
\value{
date vector
}
\description{
A function to subset data recurisvely or with a rolling window to create a valid information set. Is used as a data preparation
helper function and is called internally by forecast_univariate, forecast_multivariate, and forecast_combine.
}
