% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_formatting.R
\name{forecast_flatten}
\alias{forecast_flatten}
\title{forecast_flatten
A function to extract point estimates for each model at each date in a list output
from forecast_univariate and converts them into a matrix for forecast combination routines
or accuracy calculations}
\usage{
forecast_flatten(forecasts)
}
\arguments{
\item{forecasts}{list: list of data.frames of forecasts from forecast_univariate}
}
\value{
data.frame of forecast point estiamtes
}
\description{
forecast_flatten
A function to extract point estimates for each model at each date in a list output
from forecast_univariate and converts them into a matrix for forecast combination routines
or accuracy calculations
}
