% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_multivariate.R
\name{forecast_multivariate}
\alias{forecast_multivariate}
\title{forecast_multivariate}
\usage{
forecast_multivariate(
  Data,
  forecast.dates,
  target,
  method,
  rolling.window = NA,
  freq,
  horizon
)
}
\arguments{
\item{Data}{data.frame: data frame of target variable, exogenous variables, and observed date (named 'date')}

\item{forecast.dates}{date: dates forecasts are created}

\item{target}{string: column name in \code{Data} of variable to forecast}

\item{method}{string or vector: methods to use; 'var', 'ols', 'ridge', 'lasso', 'elastic', 'RF', 'GBM', 'NN'}

\item{rolling.window}{int: size of rolling window, NA if expanding window is used}

\item{freq}{string: time series frequency; day, week, month, quarter, year}

\item{horizon}{int: number of periods into the future to forecast}
}
\value{
data.frame with a date column and one column per forecast method selected
}
\description{
A function to estimate multivariate forecasts out-of-sample. Methods available include:
vector auto-regression, linear regression, lasso regression, ridge regression, elastic net,
random forest, tree-based gradient boosting machine, and single-layer neural network.
The function will take in a data frame of the target variable, exogenous variables, and a 'date' column,
while outputting a data frame with a date column and one column per forecast method selected.
}
