% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.formulas.R
\name{h.formulas}
\alias{h.formulas}
\title{Returning the number of candidate models.}
\usage{
h.formulas(y, df, type = "f")
}
\arguments{
\item{y}{Character : indicating the response column name.}

\item{df}{Data.frame.}

\item{type}{'f', 'fs' or 's' denoting respectively first order, first and second order, second order formulas.}
}
\value{
The number of candidate models.
}
\description{
Returning the number of candidate models.
}
