% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{Information criterion}
\usage{
ic(expr, df, family, penalty = 2)
}
\arguments{
\item{expr}{GLM formula.}

\item{df}{Data.frame.}

\item{family}{Object : a description of the error distribution.}

\item{penalty}{numeric specifying the ‘weight’ of the equivalent degrees of freedom. (penalty = 2 is the AIC and penalty = log(n) is the BIC)}
}
\value{
The Information Criterion for expr.
}
\description{
Information criterion
}
