% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{instantiate.data_impute.control_panel}
\alias{instantiate.data_impute.control_panel}
\title{Create interface to control \code{data_impute} model estimation}
\usage{
instantiate.data_impute.control_panel()
}
\value{
data_impute.control_panel
}
\description{
A function to create the data imputation method
arguments list for user manipulation.
}
