% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_combinations.R
\name{instantiate.forecast.combinations.ml.training}
\alias{instantiate.forecast.combinations.ml.training}
\title{Instantiate forecast.combinations.ml.training}
\usage{
instantiate.forecast.combinations.ml.training()
}
\value{
forecast.combinations.ml.training
}
\description{
A function to create the forecast combination technique arguments list
for user manipulation.
}
