% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_univariate.R
\name{instantiate.forecast_univariate.control_panel}
\alias{instantiate.forecast_univariate.control_panel}
\title{Instantiate forecast_univariate.control_panel}
\usage{
instantiate.forecast_univariate.control_panel()
}
\value{
forecast_univariate.control_panel
}
\description{
A function to create the univariate forecast method arguments list
for user manipulation.
}
