% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{instantiate.impute.missing.routine}
\alias{instantiate.impute.missing.routine}
\title{Instantiate impute.missing.routine}
\usage{
instantiate.impute.missing.routine()
}
\value{
instantiate.impute.missing.routine
}
\description{
A function to create the data imputation method
arguments list for user manipulation.
}
