% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_multivariate.R
\name{instantiate.multivariate.forecast.ml.training}
\alias{instantiate.multivariate.forecast.ml.training}
\title{instantiate.multivariate.forecast.ml.training}
\usage{
instantiate.multivariate.forecast.ml.training()
}
\value{
multivariate.forecast.ml.training
}
\description{
A function to create the multivariate forecast methods
arguments list for user manipulation.
}
