% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_multivariate.R
\name{instantiate.multivariate.forecast.var.training}
\alias{instantiate.multivariate.forecast.var.training}
\title{instantiate.multivariate.forecast.var.training}
\usage{
instantiate.multivariate.forecast.var.training()
}
\value{
multivariate.forecast.var.training
}
\description{
A function to create the multivariate forecast methods
arguments list for user manipulation.
}
