% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_univariate.R
\name{instantiate.univariate.forecast.training}
\alias{instantiate.univariate.forecast.training}
\title{instantiate.univariate.forecast.training}
\usage{
instantiate.univariate.forecast.training()
}
\value{
univariate.forecast.training
}
\description{
A function to create the univariate forecast method arguments list
for user manipulation.
}
