% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.formulas.R
\name{l.formulas}
\alias{l.formulas}
\title{Produce all possible formulas for a given data.frame}
\usage{
l.formulas(y, df, type = "f")
}
\arguments{
\item{y}{Character : indicating the response column name.}

\item{df}{Data.frame.}

\item{type}{'f', 'fs' or 's' denoting respectively first order, first and second order, second order formulas.}
}
\value{
All possible formulas for a given data.frame.
}
\description{
Produce all possible formulas for a given data.frame
}
