% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.ic.R
\name{l.ic}
\alias{l.ic}
\title{List - Information criterion}
\usage{
l.ic(exprs, df, family, penalty = 2)
}
\arguments{
\item{exprs}{List of GLM formulas.}

\item{df}{Data.frame.}

\item{family}{Object : a description of the error distribution.}

\item{penalty}{numeric specifying the ‘weight’ of the equivalent degrees of freedom. (penalty = 2 is the AIC and penalty = log(n) is the BIC)}
}
\value{
The Information Criterion for all formulas in 'exprs'.
}
\description{
List - Information criterion
}
