% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.cv.R
\name{ml.cv}
\alias{ml.cv}
\title{Multicore l.cv for heavy tasks}
\usage{
ml.cv(
  exprs,
  df,
  family,
  thresh = FALSE,
  statistic = "Accuracy",
  number = 10,
  repeats = 10,
  cl = detectCores()
)
}
\arguments{
\item{exprs}{List of GLM formulas.}

\item{df}{Data.frame.}

\item{family}{Object : a description of the error distribution.}

\item{thresh}{If TRUE, an optimal threshold is found based on the 'statistic' criterion.}

\item{statistic}{A character indicating which statistic to maximize. Default is "Accuracy". You can use c('Sensitivity', 'Specificity', 'Pos Pred Value', 'Neg Pred Value', 'Precision', 'Accuracy', 'Balanced Accuracy', 'F1', 'Detection Rate', 'Detection Prevalence', 'Prevalence', 'Recall', 'Kappa', 'J', 'Dist').}

\item{number}{The number of folds.}

\item{repeats}{The number of complete sets of folds to compute.}

\item{cl}{Number of CPU cores on the current host.}
}
\value{
Different accuracy measures for all formulas in 'exprs'.
}
\description{
Multicore l.cv for heavy tasks
}
