% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.formulas.R
\name{ml.formulas}
\alias{ml.formulas}
\title{Multicore l.formulas for heavy tasks}
\usage{
ml.formulas(y, df, type = "f", cl = detectCores())
}
\arguments{
\item{y}{Character : indicating the response column name.}

\item{df}{Data.frame.}

\item{type}{'f', 'fs' or 's' denoting respectively first order, first and second order, second order formulas.}

\item{cl}{Number of CPU cores on the current host.}
}
\value{
All possible formulas for a given data.frame.
}
\description{
Multicore l.formulas for heavy tasks
}
