% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.ic.R
\name{ml.ic}
\alias{ml.ic}
\title{Multicore l.ic for heavy tasks}
\usage{
ml.ic(exprs, df, family, penalty = 2, cl = detectCores())
}
\arguments{
\item{exprs}{List of GLM formulas.}

\item{df}{Data.frame.}

\item{family}{Object : a description of the error distribution.}

\item{penalty}{numeric specifying the ‘weight’ of the equivalent degrees of freedom. (penalty = 2 is the AIC and penalty = log(n) is the BIC)}

\item{cl}{Number of CPU cores on the current host.}
}
\value{
The Information Criterion for all formulas in 'exprs'.
}
\description{
Multicore l.ic for heavy tasks
}
