% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{n.lag}
\alias{n.lag}
\title{Create n lags}
\usage{
n.lag(Data, lags, variables = NULL)
}
\arguments{
\item{Data}{data.frame: data frame of variables to lag and a 'date' column}

\item{lags}{int: number of lags to create}

\item{variables}{string: vector of variable names to lag, default is all non-date variables}
}
\value{
data.frame
}
\description{
A function to create 1 through n lags of a set of variables. Is used as a data preparation
helper function and is called internally by forecast_univariate, forecast_multivariate, and forecast_combine.
}
