% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_combinations.R
\name{NBest}
\alias{NBest}
\title{NBest}
\usage{
NBest(forecasts, n.max, window = NA)
}
\arguments{
\item{forecasts}{data.frame: a data frame of forecasts to combine, assumes one column named "observed"}

\item{n.max}{int: maximum number of forecasts to select}

\item{window}{int: size of rolling window to evaluate forecast error over, use entire period if NA}
}
\value{
data.frame with n columns of the historically best forecasts
}
\description{
A function to subset the n-best forecasts;
assumes column named observed.
}
