% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{standardize}
\alias{standardize}
\title{Standardize variables (mean 0, variance 1)}
\usage{
standardize(X)
}
\arguments{
\item{X}{numeric: vector to be winsorized or trimmed}

\item{bounds}{double: vector of winsorizing minimum and maximum bounds, c(min percentile, max percentile)}

\item{replacement}{boolean: if TRUE then replace outliers with NA instead of winsorizing bound}
}
\value{
numeric vector of winsorized or trimmed values
}
\description{
Standardize variables (mean 0, variance 1)
}
