% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_combinations.R
\name{winsorize}
\alias{winsorize}
\title{winsorize}
\usage{
winsorize(X, bounds)
}
\arguments{
\item{X}{numeric: numeric data to be winsorized}

\item{bounds}{numeric: upper and lower winsorizing bounds}
}
\value{
vector of winsorized  X
}
\description{
A function to winsorize the cross-section of forecasts
}
