% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_chart.R
\name{chart_forecast}
\alias{chart_forecast}
\title{Chart forecasts}
\usage{
chart_forecast(Data, Title, Ylab, Freq, zeroline = FALSE)
}
\arguments{
\item{Data}{data.frame: oos.forecast object}

\item{Title}{string: chart title}

\item{Ylab}{string: y-axis label}

\item{Freq}{string: frequency (acts as sub-title)}

\item{zeroline}{boolean: if TRUE then add a horizontal line at zero}
}
\value{
ggplot2 chart
}
\description{
Chart forecasts
}
