% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_metrics.R
\name{forecast_accuracy}
\alias{forecast_accuracy}
\title{Forecast accuracy}
\usage{
forecast_accuracy(Data)
}
\arguments{
\item{Data}{data.frame: data frame of forecasts, model names, and dates}
}
\value{
data.frame of numeric error results
}
\description{
A function to calculate various error loss functions. Options include:
MSE, RMSE, MAE, and  MAPE. The default is MSE loss.
}
