% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_metrics.R
\name{forecast_accuracy}
\alias{forecast_accuracy}
\title{Forecast accuracy}
\usage{
forecast_accuracy(forecast, observed, metric = "MSE")
}
\arguments{
\item{forecast}{numeric: vector of forecasted values}

\item{observed}{numeric: vector of observed values}

\item{metric}{string: loss function}
}
\value{
numeric test result
}
\description{
A function to calculate various error loss functions. Options include:
MSE, RMSE, MAE, and  MAPE. The default is MSE loss.
}
