% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_combinations.R
\name{forecast_combine}
\alias{forecast_combine}
\title{forecast_combine}
\usage{
forecast_combine(
  forecasts,
  method = "unform",
  n.max = NULL,
  window = NA,
  trim = c(0.5, 0.95),
  burn.in = 1
)
}
\description{
A function to combine forecasts out-of-sample. Methods available include:
uniform weights, median forecast, trimmed (winsorized) mean, n-best,
ridge regression, lasso regression, elastic net, peLASSO,
random forest, tree-based gradient boosting machine, and single-layer neural network.
The function will take in a data frame of forecasts, observed values (named 'observed', when the method requires it),
and the observation date (i.e. the foretasted date, named 'date'), while outputting a data frame
with a date column and one column per combination method selected.
}
\details{
@param forecasts   data.frame: data frame of forecasted values to combine, assumes \code{date} and \code{observed} columns, but `observed' is not necessary for all methods
@param method      string or vector: the method to use; 'uniform', 'median', 'trimmed.mean', 'n.best', 'peLasso', 'lasso', 'ridge', 'elastic', 'RF', 'GBM', 'NN'
@param n.max       int: maximum number of forecasts to select
@param window      int: size of rolling window to evaluate forecast error over, use entire period if NA
@param trim        numeric: a two element vector with the winsorizing bounds for the trimmed mean method; c(min, max)
@param burn.in     int: the number of periods to use in the first model estimation

@return  data.frame with a date column and one column per forecast method selected

@export
}
