% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_metrics.R
\name{forecast_comparison}
\alias{forecast_comparison}
\title{forecast_comparison}
\usage{
forecast_comparison(
  baseline.forecast,
  alternative.forecast,
  observed,
  test = "ER",
  loss = "MSE",
  horizon = NULL
)
}
\arguments{
\item{baseline.forecast}{numeric: vector of baseline (null hypothesis) forecasts}

\item{alternative.forecast}{numeric: vector of alternative forecasts}

\item{observed}{numeric: vector of observed values}

\item{test}{string: which test to use; ER = error ratio, DM = Diebold-Mariano, CM = Clark and West}

\item{loss}{string: error loss function to use if creating forecast error ratio}

\item{horizon}{int: horizon of forecasts being compared in DM and CW tests}
}
\value{
numeric test result
}
\description{
A function to compare two forecasts. Options include: simple forecast error ratios,
Diebold-Mariano test, and Clark and West test for nested models
}
