% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_univariate.R
\name{forecast_univariate}
\alias{forecast_univariate}
\title{Forecast with univariate models}
\usage{
forecast_univariate(
  Data,
  forecast.dates,
  methods,
  periods,
  rolling.window = NA,
  freq,
  recursive = TRUE
)
}
\arguments{
\item{Data}{data.frame: data frame of variable to forecast and a date column}

\item{forecast.dates}{date: dates forecasts are created}

\item{methods}{string or vector: models to estimate forecasts}

\item{periods}{int: number of periods to forecast}

\item{rolling.window}{int: size of rolling window, NA if expanding window is used}

\item{freq}{string: time series frequency; day, week, month, quarter, year}

\item{recursive}{boolean: use sequential one-step-ahead forecast if TRUE, use direct projections if FALSE}
}
\value{
data.frame with a date column and one column per forecast method selected
}
\description{
A function to estimate univariate forecasts out-of-sample. Methods available include all forecast
methods from the \code{forecast} package. The function will take in a data frame of the target variable,
and a observation date column (named 'date'), while outputting a data frame with a date column and
one column per forecast method selected.
}
