% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_combinations.R
\name{instantiate.forecast_combinations.control_panel}
\alias{instantiate.forecast_combinations.control_panel}
\title{Create interface to control \code{forecast_combine} model estimation}
\usage{
instantiate.forecast_combinations.control_panel(
  covariates = NULL,
  rolling.window = NULL,
  horizon = NULL
)
}
\arguments{
\item{covariates}{int: the number of features that will go into the model}

\item{rolling.window}{int: size of rolling window, NA if expanding window is used}

\item{horizon}{int: number of periods into the future to forecast}
}
\value{
forecast_combinations.control_panel
}
\description{
A function to create the forecast combination technique arguments list
for user manipulation.
}
