% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_combinations.R
\name{instantiate.forecast_combinations.control_panel}
\alias{instantiate.forecast_combinations.control_panel}
\title{Create interface to control \code{forecast_combine} model estimation}
\usage{
instantiate.forecast_combinations.control_panel()
}
\value{
forecast_combinations.control_panel
}
\description{
A function to create the forecast combination technique arguments list
for user manipulation.
}
