% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_multivariate.R
\name{instantiate.forecast_multivariate.ml.control_panel}
\alias{instantiate.forecast_multivariate.ml.control_panel}
\title{Create interface to control \code{forecast_multivariate} ML estimation}
\usage{
instantiate.forecast_multivariate.ml.control_panel()
}
\value{
forecast_multivariate.ml.control_panel
}
\description{
A function to create the multivariate forecast methods
arguments list for user manipulation.
}
