% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{standardize}
\alias{standardize}
\title{Standardize variables (mean 0, variance 1)}
\usage{
standardize(X)
}
\arguments{
\item{X}{numeric: vector to be standardized}
}
\value{
numeric vector of standardized values
}
\description{
Standardize variables (mean 0, variance 1)
}
