% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize or trim variables}
\usage{
winsorize(X, bounds, trim = FALSE)
}
\arguments{
\item{X}{numeric: vector to be winsorized or trimmed}

\item{bounds}{double: vector of winsorizing minimum and maximum bounds, c(min percentile, max percentile)}

\item{trim}{boolean: if TRUE then replace outliers with NA instead of winsorizing bound}
}
\value{
numeric vector of winsorized or trimmed values
}
\description{
Winsorize or trim variables
}
