% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R, R/forecast_combinations.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize or trim variables}
\usage{
winsorize(X, bounds)

winsorize(X, bounds)
}
\arguments{
\item{X}{numeric: numeric data to be winsorized}

\item{bounds}{numeric: upper and lower winsorizing bounds}

\item{replacement}{boolean: if TRUE then replace outliers with NA instead of winsorizing bound}
}
\value{
numeric vector of winsorized or trimmed values

vector of winsorized  X
}
\description{
A function to winsorize the cross-section of forecasts
}
